change .jsx to Capital, add Rules.jsx

+++ js/Images.jsx
@@ -1,15 +1,10 @@
-var React = require('react');
-var ReactDOM = require('react-dom');
+import React from 'react'
+import ReactDOM from 'react-dom'
 
-var Images=React.createClass({
-  imgClick:function(item,index){
-    if (this.props.sendMeToUser) {
-      //点击后改变results里面的数据，也许用redux会比较方便？这儿其实比较简单。
-      this.props.sendMeToUser(item)
-    };
-    if (this.props.deleteMe) {
-      this.props.deleteMe(item)
-    }
+var Images = React.createClass({
+  imgClick: function(item,index){
+    if(this.props.sendMeToUser) {   this.props.sendMeToUser(item)   }
+    if(this.props.deleteMe)     {   this.props.deleteMe(item)       }
   },
   render: function(){
      const lineBreak = this.props.lineBreak? this.props.lineBreak : 100
@@ -30,6 +25,7 @@ var Images=React.createClass({
        </div>
      );
   }
-});
+})
 
-module.exports = Images
+module.exports = Images;
+
